Add async support for app uninstaller.
