+++ PlayCover/Views/Sidebar Views/AppLibraryView.swift
@@ -9,6 +9,7 @@ struct AppLibraryView: View {
     @EnvironmentObject var appsVM: AppsVM
     @EnvironmentObject var installVM: InstallVM
     @EnvironmentObject var downloadVM: DownloadVM
+    @EnvironmentObject var uninstallVM: UninstallVM
 
     @Binding var selectedBackgroundColor: Color
     @Binding var selectedTextColor: Color
@@ -33,6 +34,7 @@ struct AppLibraryView: View {
                                             selected: $selected,
                                             app: app,
                                             isList: isList)
+                                .environmentObject(uninstallVM)
                             }
                         }
                         .padding()
@@ -44,6 +46,7 @@ struct AppLibraryView: View {
                                             selected: $selected,
                                             app: app,
                                             isList: isList)
+                                .environmentObject(uninstallVM)
                             }
                             Spacer()
                         }
@@ -111,6 +114,7 @@ struct AppLibraryView: View {
             }
         }
         .searchable(text: $searchString, placement: .toolbar)
+        .disabled(uninstallVM.inProgress)
         .onChange(of: searchString, perform: { value in
             appsVM.searchText = value
             appsVM.fetchApps()
