+++ PlayCover/en.lproj/Localizable.strings
@@ -90,6 +90,8 @@
 "playapp.install.signing" = "Signing app...";
 "playapp.install.addToLib" = "Adding app to library...";
 "playapp.install.copy" = "Copying app...";
+"playapp.uninstall.begin" = "Uninstalling app...";
+"playapp.uninstall.finished" = "Finished";
 "playapp.progress.finished" = "Finished";
 "playapp.progress.failed" = "Failed";
 "playapp.progress.canceled" = "Canceled";
